Add small refactor to utils
