+++ actions/utils/poll.ts
@@ -2,10 +2,9 @@ import {
   ConditionalOrderFactory,
   DEFAULT_CONDITIONAL_ORDER_REGSTRY,
   PollResult,
-  PollResultCode,
 } from "@cowprotocol/cow-sdk";
 
-import { PollingParams } from "./model";
+import { PollingParams } from "../model";
 
 const ordersFactory = new ConditionalOrderFactory(
   DEFAULT_CONDITIONAL_ORDER_REGSTRY
@@ -16,7 +15,7 @@ export async function pollConditionalOrder(
 ): Promise<PollResult | undefined> {
   const { owner, chainId, conditionalOrderParams, provider } = params;
 
-  const order = ordersFactory.fromParams(params.conditionalOrderParams);
+  const order = ordersFactory.fromParams(conditionalOrderParams);
 
   if (!order) {
     return undefined;
