+++ actions/utils/provider.ts
@@ -1,0 +1,47 @@
+import { Context } from "@tenderly/actions";
+
+import { ethers } from "ethers";
+import { ConnectionInfo, Logger } from "ethers/lib/utils";
+
+import { SupportedChainId } from "@cowprotocol/cow-sdk";
+import assert = require("assert");
+
+export async function getProvider(
+  context: Context,
+  chainId: SupportedChainId
+): Promise<ethers.providers.Provider> {
+  Logger.setLogLevel(Logger.levels.DEBUG);
+
+  const url = await getSecret(`NODE_URL_${chainId}`, context);
+  const user = await getSecret(`NODE_USER_${chainId}`, context).catch(
+    () => undefined
+  );
+  const password = await getSecret(`NODE_PASSWORD_${chainId}`, context).catch(
+    () => undefined
+  );
+  const providerConfig: ConnectionInfo =
+    user && password
+      ? {
+          url,
+          // TODO: This is a hack to make it work for HTTP endpoints (while we don't have a HTTPS one for Gnosis Chain), however I will delete once we have it
+          headers: {
+            Authorization: getAuthHeader({ user, password }),
+          },
+          // user: await getSecret(`NODE_USER_${network}`, context),
+          // password: await getSecret(`NODE_PASSWORD_${network}`, context),
+        }
+      : { url };
+
+  return new ethers.providers.JsonRpcProvider(providerConfig);
+}
+
+async function getSecret(key: string, context: Context): Promise<string> {
+  const value = await context.secrets.get(key);
+  assert(value, `${key} secret is required`);
+
+  return value;
+}
+
+function getAuthHeader({ user, password }: { user: string; password: string }) {
+  return "Basic " + Buffer.from(`${user}:${password}`).toString("base64");
+}
